Update PHP 8.0 Test and Fixtures
