+++ tests/Fixture/PHP80000/ArgumentMixedTypeHint.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace MockeryTest\Fixture\PHP80000;
+
+class ArgumentMixedTypeHint
+{
+    public function foo(mixed $foo)
+    {
+    }
+}
