+++ tests/Fixture/PHP80000/ArgumentParentTypeHint.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace MockeryTest\Fixture\PHP80000;
+
+class ArgumentParentTypeHint extends \stdClass
+{
+    public function foo(parent $foo)
+    {
+    }
+}
