+++ tests/Fixture/PHP80000/ArgumentUnionTypeHint.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace MockeryTest\Fixture\PHP80000;
+
+class ArgumentUnionTypeHint
+{
+    public function foo(string|array|self $foo)
+    {
+    }
+}
