+++ tests/Fixture/PHP80000/ArgumentUnionTypeHintWithNull.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace MockeryTest\Fixture\PHP80000;
+
+class ArgumentUnionTypeHintWithNull
+{
+    public function foo(string|array|null $foo)
+    {
+    }
+}
