+++ tests/Fixture/PHP80000/ImplementsIterator.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace MockeryTest\Fixture\PHP80000;
+
+class ImplementsIterator implements \Iterator
+{
+    public function rewind(): void
+    {
+    }
+    public function current(): mixed
+    {
+    }
+    public function key(): mixed
+    {
+    }
+    public function next(): void
+    {
+    }
+    public function valid(): bool
+    {
+    }
+}
