+++ tests/Fixture/PHP80000/ReturnTypeMixedTypeHint.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace MockeryTest\Fixture\PHP80000;
+
+class ReturnTypeMixedTypeHint
+{
+    public function foo(): mixed
+    {
+    }
+}
