+++ tests/Fixture/PHP80000/ReturnTypeParentTypeHint.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace MockeryTest\Fixture\PHP80000;
+
+class ReturnTypeParentTypeHint extends \stdClass
+{
+    public function foo(): parent
+    {
+    }
+}
