+++ tests/Fixture/PHP80000/ReturnTypeUnionTypeHint.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace MockeryTest\Fixture\PHP80000;
+
+class ReturnTypeUnionTypeHint
+{
+    public function foo(): ReturnTypeMixedTypeHint|self
+    {
+    }
+}
